refactor(response): Move parsing back into a separate module
